Update patch set 2
Patch Set 2:

Hmm, I'm not sure how to do that. I thought following the same definitions as the other file systems was sufficient.

It goes through vold yes, but fuse has it's own exfat driver. Not sure about sdcardfs though. But with the fuse one I had a lot of issues, so I'm using the in-kernel driver, and without this patch I can't read or use the external sdcard.

Patch-set: 2