Update patch set 1
Patch Set 1:

> I don't, but it's not just 5.1.1 perfd, I have a 6.0 perfd that
 > does as well.

using some kernel hacking magic I found out that the problem isn't ptrace, it's readdir on "/proc" which generates the ptrace denial for /proc/PID directories which you're not allowed to access.

In fact, selinux doesn't allow perfd to access(=list using readdir) any of these directories except for a few like 1 and self.

Is there a reason for that? any other process like the shell or a self-written C application is allowed to do so.

Patch-set: 1