Copy the plain text file_contexts in out/
Android is now using a binary file_contexts instead of a plain text
one, however recovery was not updated to use it. Save the
intermediate plain text file_contexts so that it can be included
in the root of flashable zips to provide backward compatibility.
This will also provide support for old recoveries once the support
for the new format will be added.

Change-Id: I08d62bc97ff54fdf6f288722c81405cabc861f7a