Update patch set 1
Patch Set 1:

> This seems to need the make update-api to be executed too..

Yes this is needed, I dunno if there is a way to remove to update the API directly in Gerrit so that no need to make the update-api. This update only needs to be done when switching ROMs or make clobber. If you make the same build with the same OMS commits then no need to do it again.

Patch-set: 1